add CosmeticFoldingLine notion
Create cosmetic folding lines in the middle of straight edges. There is
no folding on such CosmeticFoldingLine, it is only used as a guide.
It would also be nice to draw them in a different color, but it seems
that changing a line color requires a `Part::Feature` object added to
the active document :/
cf https://forum.freecad.org/viewtopic.php?t=7335